workspace/cereja_os/telegram/requirements_telegram.txt
# Telegram Bot Dependencies - Cereja OS
# python-telegram-bot alternative: aiogram v3 (recommended for modern async)

# Core bot framework
aiogram>=3.13.0,<4.0.0

# HTTP server for webhook (aiohttp)
aiohttp>=3.9.0,<4.0.0

# Environment variables management
python-dotenv>=1.0.0,<2.0.0

# Async support
asyncio-throttle>=1.0.2,<2.0.0

# Optional: HTTP client for Orquestrador API calls
httpx>=0.27.0,<1.0.0

# Optional: Redis for session/state storage (multi-tenant)
redis>=5.0.0,<6.0.0

# Optional: Structured logging
python-json-logger>=2.0.0,<3.0.0
